+++ packages/fighting-design/_hooks/use-props/index.ts
@@ -32,7 +32,7 @@ export const useProps = <T extends object>(prop: T): UsePropsReturn => {
       if (isString(item)) {
         result[item] = toRef(prop, item as never)
       } else if (isObject(item)) {
-      /** 如果是对象类型 */
+        /** 如果是对象类型 */
         /**
          * 如果是 object 类型
          *